+++ lib/features/browse/components/creator_card.dart
@@ -1,0 +1,74 @@
+import 'package:floaty/features/api/models/definitions.dart';
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+class CreatorCard extends StatefulWidget {
+  final CreatorModelV3 creator;
+
+  const CreatorCard(this.creator, {super.key});
+
+  @override
+  CreatorCardState createState() => CreatorCardState();
+}
+
+class CreatorCardState extends State<CreatorCard> {
+  @override
+  Widget build(BuildContext context) {
+    return LayoutBuilder(
+      builder: (context, constraints) {
+        return Material(
+          color: Theme.of(context).colorScheme.surfaceContainer,
+          borderRadius: BorderRadius.circular(8),
+          child: InkWell(
+            borderRadius: BorderRadius.circular(8),
+            onTap: () => context.push('/channel/${widget.creator.urlname}'),
+            child: Column(
+              children: [
+                Expanded(
+                  flex: 4,
+                  child: Padding(
+                    padding: const EdgeInsets.fromLTRB(12, 12, 12, 8),
+                    child: AspectRatio(
+                      aspectRatio: 1,
+                      child: Container(
+                        decoration: BoxDecoration(
+                          shape: BoxShape.circle,
+                          image: DecorationImage(
+                            image: NetworkImage(
+                              widget.creator.icon?.path ?? '',
+                            ),
+                            fit: BoxFit.cover,
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+                Expanded(
+                  flex: 1,
+                  child: Padding(
+                    padding: const EdgeInsets.fromLTRB(8, 0, 8, 8),
+                    child: Center(
+                      child: Text(
+                        widget.creator.title ?? '',
+                        textAlign: TextAlign.center,
+                        style: const TextStyle(
+                          fontSize: 14,
+                          fontWeight: FontWeight.bold,
+                        ),
+                        maxLines: 2,
+                        overflow: TextOverflow.ellipsis,
+                      ),
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        );
+      },
+    );
+  }
+}
+
+enum CardState { initial, expanded }
