+++ lib/features/browse/repositories/browse_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:floaty/features/api/models/definitions.dart';
 import 'package:floaty/features/api/repositories/fpapi.dart';
-import 'package:floaty/shared/components/root.dart';
+import 'package:floaty/features/router/views/root_layout.dart';
 import 'package:flutter/material.dart';
 import 'dart:async';
