+++ lib/features/channel/components/stat_column.dart
@@ -1,0 +1,48 @@
+import 'package:flutter/material.dart';
+
+class StatColumn extends StatelessWidget {
+  final String value;
+  final String label;
+
+  const StatColumn({super.key, required this.value, required this.label});
+
+  @override
+  Widget build(BuildContext context) {
+    return LayoutBuilder(
+      builder: (context, constraints) {
+        final maxWidth = constraints.maxWidth;
+        final double valueFontSize = (maxWidth * 0.15).clamp(16.0, 24.0);
+        final double labelFontSize = (maxWidth * 0.08).clamp(12.0, 14.0);
+
+        return Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            FittedBox(
+              fit: BoxFit.scaleDown,
+              child: Text(
+                value,
+                style: TextStyle(
+                  fontSize: valueFontSize,
+                  fontWeight: FontWeight.bold,
+                ),
+                textAlign: TextAlign.center,
+              ),
+            ),
+            const SizedBox(height: 4.0),
+            FittedBox(
+              fit: BoxFit.scaleDown,
+              child: Text(
+                label,
+                style: TextStyle(
+                  fontSize: labelFontSize,
+                  color: Theme.of(context).textTheme.titleMedium?.color,
+                ),
+                textAlign: TextAlign.center,
+              ),
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
