+++ lib/features/post/views/post_screen.dart
@@ -1,8 +1,13 @@
-import 'package:floaty/shared/components/elements.dart';
+import 'package:floaty/features/post/components/blog_post_card.dart';
+import 'package:floaty/features/post/components/comment_holder.dart';
+import 'package:floaty/features/post/components/expandable_description.dart';
+import 'package:floaty/features/post/components/state_card.dart';
+
+import 'package:floaty/shared/views/error_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:floaty/features/post/repositories/post_provider.dart';
-import 'package:floaty/shared/components/root.dart';
+import 'package:floaty/features/router/views/root_layout.dart';
 import 'package:floaty/features/api/repositories/fpapi.dart';
 import 'package:floaty/features/api/models/definitions.dart';
 import 'dart:async';
